Update on the main.py file -> It is now Game.py, which serves to manage the game. Its initialization is handled by app.py, where a better graphical interface will be implemented in the future.
Update on the word.py file -> It is now WordleCore.py. It was changed to function as an object, allowing it to work better with the database in a more organized and comprehensible manner, along with new methods to make this possible.

Update on Game.py -> Previously, the bot was picking up all messages, whether from the bot itself or from the user, and responding incorrectly. Now, there is a filter to respond only to user messages.

Update on Data.py -> Some methods had to be altered to work with the new updates in the Wordle class.

Addition: Interface.py -> It was created with the intention of providing a graphical interface for better bot configuration; however, it is still in the development phase. >A lot of improvement is needed here<

Addition: app.py -> It is now responsible for initializing the bot.

Addition: appserver.py -> Without a graphical interface, it assists in initializing the bot through a server where a graphical interface is not possible.